Fixed type

+++ restcomet.cpp
@@ -407,11 +407,11 @@ void restcomet::CheckClientEvents( http_client& client )
 
 string restcomet::TrimStr(const string& src, const string& c )
 {
-	int p2 = src.find_last_not_of( c );
+	size_t p2 = src.find_last_not_of( c );
 	if ( p2 == std::string::npos )
 		return std::string();
 
-	int p1 = src.find_first_not_of( c );
+	size_t p1 = src.find_first_not_of( c );
 	if ( p1 == std::string::npos )
 		p1 = 0;
